feat: move package ViewResolvers
