--- nextstep-mvc/src/main/java/nextstep/mvc/tobe/view/ViewResolver.java
@@ -1,7 +1,0 @@
-package nextstep.mvc.tobe.view;
-
-public interface ViewResolver {
-    boolean support(ViewType viewType);
-
-    View resolve(String viewName);
-}
