+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/view/viewresolver/JsonViewResolver.java
@@ -1,4 +1,8 @@
-package nextstep.mvc.tobe.view;
+package nextstep.mvc.tobe.view.viewresolver;
+
+import nextstep.mvc.tobe.view.JsonView;
+import nextstep.mvc.tobe.view.View;
+import nextstep.mvc.tobe.view.ViewType;
 
 public class JsonViewResolver implements ViewResolver {
