+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/view/viewresolver/JspViewResolver.java
@@ -1,4 +1,8 @@
-package nextstep.mvc.tobe.view;
+package nextstep.mvc.tobe.view.viewresolver;
+
+import nextstep.mvc.tobe.view.JspView;
+import nextstep.mvc.tobe.view.View;
+import nextstep.mvc.tobe.view.ViewType;
 
 public class JspViewResolver implements ViewResolver {
